FACES-1728 NullPointerException thrown by UploadedFileCleanupListener on WebLogic during session expiration

+++ demos/portal/primefaces5-users-portlet/src/main/java/com/liferay/faces/demos/servlet/UploadedFileCleanupListener.java
@@ -49,12 +49,17 @@ public class UploadedFileCleanupListener implements HttpSessionListener {
 			String parent = UploadedFileUtil.getTempDir();
 			String folderName = sessionId;
 			File folder = new File(parent, folderName);
-			File[] fileList = folder.listFiles().clone();
+			File[] fileList = folder.listFiles();
 
-			for (File file : fileList) {
+			if (fileList != null) {
 
-				if (file.getName().startsWith(UploadedFileUtil.USER_PORTRAIT)) {
-					file.delete();
+				fileList = fileList.clone();
+
+				for (File file : fileList) {
+
+					if (file.getName().startsWith(UploadedFileUtil.USER_PORTRAIT)) {
+						file.delete();
+					}
 				}
 			}
